split out gauge.js
