+++ script/main.js
@@ -3,6 +3,7 @@ import {fromEvent, combineLatest, interval} from 'rxjs';
 import {map, startWith, debounce} from 'rxjs/operators';
 import {$, $$, calculateRating, arithmeticMean, NBSP, numberFormatter} from './util.js';
 import {weights, scoring} from './metrics.js';
+import { updateGauge } from './gauge.js';
 
 
 function main(weights, container) {
@@ -202,16 +203,7 @@ function main(weights, container) {
       })
     )
     .subscribe(score => {
-      const wrapper = container.$('.lh-gauge__wrapper');
-      wrapper.className = 'lh-gauge__wrapper'; // clear any other labels already set
-      wrapper.classList.add(`lh-gauge__wrapper--${calculateRating(score / 100)}`);
-
-      const gaugeArc = container.$('.lh-gauge-arc');
-      gaugeArc.style.strokeDasharray = `${(score / 100) * 352} 352`;
-
-      const scoreOutOf100 = Math.round(score);
-      const percentageEl = container.$('.lh-gauge__percentage');
-      percentageEl.textContent = scoreOutOf100.toString();
+      updateGauge(container, score);
     });
 }
